Refactor onboarding_page.dart into view/onboarding_view.dart     Set up ViewMoodel (onboarding_viewmodel.dart) to the Onboarding Screens
