--- lib/ui/view/on_boarding/onboading_view.dart
@@ -1,1 +1,0 @@
-import 'package:flutter/material.dart';
