--- lib/ui/view/on_boarding/onboarding_viewmodel.dart
@@ -1,1 +1,0 @@
-import 'package:flutter/material.dart';
