+++ lib/ui/view/onboarding/onboarding_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class OnboardingViewModel extends BaseViewModel {}
